Move TimingFunction into TransitionSettings, other minor doc changes
